BLS CoSi protocol refactoring
This cleans some part of the main protocol
