+++ blsftcosi/protocol/protocol.go
@@ -14,7 +14,6 @@ import (
 	"github.com/dedis/kyber/sign/cosi"
 	"github.com/dedis/onet"
 	"github.com/dedis/onet/log"
-	"github.com/dedis/onet/network"
 )
 
 // VerificationFn is called on every node. Where msg is the message that is
@@ -112,7 +111,9 @@ func (p *BlsFtCosi) SetNbrSubTree(nbr int) error {
 func (p *BlsFtCosi) Shutdown() error {
 	p.stoppedOnce.Do(func() {
 		for _, subFtCosi := range p.subProtocols {
-			subFtCosi.HandleStop(StructStop{subFtCosi.TreeNode(), Stop{}})
+			// we're stopping the root thus it will stop the children
+			// by itself
+			subFtCosi.Shutdown()
 		}
 		close(p.startChan)
 		close(p.FinalSignature)
@@ -180,7 +181,7 @@ func (p *BlsFtCosi) Dispatch() error {
 	}
 
 	// generate root signature
-	signaturePoint, finalMask, err := generateSignature(p.suite, p.TreeNodeInstance, p.Roster().Publics(), p.Private(), responses, p.Msg, verificationOk)
+	signaturePoint, finalMask, err := p.generateSignature(responses, verificationOk)
 	if err != nil {
 		return err
 	}
@@ -243,6 +244,14 @@ func (p *BlsFtCosi) checkIntegrity() error {
 	return nil
 }
 
+func (p *BlsFtCosi) checkFailureThreshold(numFailure int) bool {
+	if numFailure == 0 {
+		return false
+	}
+
+	return numFailure > len(p.Roster().List)-p.Threshold
+}
+
 // startSubProtocol creates, parametrize and starts a subprotocol on a given tree
 // and returns the started protocol.
 func (p *BlsFtCosi) startSubProtocol(tree *onet.Tree) (*SubBlsFtCosi, error) {
@@ -273,50 +282,143 @@ func (p *BlsFtCosi) startSubProtocol(tree *onet.Tree) (*SubBlsFtCosi, error) {
 	return cosiSubProtocol, err
 }
 
+// Collect signatures from each sub-leader, restart whereever sub-leaders fail to respond.
+// The collected signatures are already aggregated for a particular group
+func (p *BlsFtCosi) collectSignatures() (ResponseMap, error) {
+	responsesChan := make(chan StructResponse, len(p.subProtocols))
+	errChan := make(chan error, len(p.subProtocols))
+	closeChan := make(chan bool)
+
+	for i, subProtocol := range p.subProtocols {
+		go func(i int, subProtocol *SubBlsFtCosi) {
+			timeout := time.After(p.Timeout)
+			for {
+				select {
+				case <-closeChan:
+					// quick answer/failure
+					return
+				case <-subProtocol.subleaderNotResponding:
+					subleaderID := p.subTrees[i].Root.Children[0].RosterIndex
+					log.Lvlf2("(subprotocol %v) subleader with id %d failed, restarting subprotocol", i, subleaderID)
+
+					// generate new tree by adding the current subleader to the end of the
+					// leafs and taking the first leaf for the new subleader.
+					nodes := []int{p.subTrees[i].Root.RosterIndex}
+					for _, child := range p.subTrees[i].Root.Children[0].Children {
+						nodes = append(nodes, child.RosterIndex)
+					}
+
+					if len(nodes) < 2 || subleaderID > nodes[1] {
+						errChan <- fmt.Errorf("(subprotocol %v) failed with every subleader, ignoring this subtree",
+							i)
+						return
+					}
+					nodes = append(nodes, subleaderID)
+
+					var err error
+					p.subTrees[i], err = genSubtree(p.subTrees[i].Roster, nodes)
+					if err != nil {
+						errChan <- fmt.Errorf("(subprotocol %v) error in tree generation: %v", i, err)
+						return
+					}
+
+					// restart subprotocol
+					// send stop signal to old protocol
+					subProtocol.HandleStop(StructStop{subProtocol.TreeNode(), Stop{}})
+					subProtocol, err = p.startSubProtocol(p.subTrees[i])
+					if err != nil {
+						errChan <- fmt.Errorf("(subprotocol %v) error in restarting of subprotocol: %s", i, err)
+						return
+					}
+
+					p.subProtocols[i] = subProtocol
+				case response := <-subProtocol.subResponse:
+					responsesChan <- response
+					return
+				case <-timeout:
+					// This should never happen, as the subProto should return before that
+					// timeout, even if it didn't receive enough responses.
+					errChan <- fmt.Errorf("timeout should not happen while waiting for response: %d", i)
+					return
+				}
+			}
+		}(i, subProtocol)
+	}
+
+	// handle answers from all parallel threads
+	responseMap := make(ResponseMap)
+	numSignature := 0
+	numFailure := 0
+	for len(p.subProtocols) > 0 && numSignature < p.Threshold-1 && !p.checkFailureThreshold(numFailure) {
+		select {
+		case res := <-responsesChan:
+			mask, err := cosi.NewMask(p.suite.(cosi.Suite), p.Roster().Publics(), nil)
+			if err != nil {
+				return nil, err
+			}
+			err = mask.SetMask(res.Mask)
+			if err != nil {
+				return nil, err
+			}
+
+			numSignature += mask.CountEnabled()
+			numFailure += res.SubtreeCount() - mask.CountEnabled()
+			responseMap[res.ID] = &res.Response
+		case err := <-errChan:
+			err = fmt.Errorf("error in getting responses: %s", err)
+			return nil, err
+		case <-time.After(p.Timeout):
+			return nil, fmt.Errorf("not enough replies from nodes at timeout %v "+
+				"for Threshold %d, got %d responses for %d requests", p.Timeout,
+				p.Threshold, numSignature, len(p.Roster().List)-1)
+		}
+	}
+
+	// Force to stop pending subprotocols in case of quick answer/failure scenario
+	close(closeChan)
+
+	if p.checkFailureThreshold(numFailure) {
+		return nil, fmt.Errorf("too many refusals (got %d), the threshold of %d cannot be achieved",
+			numFailure, p.Threshold)
+	}
+
+	return responseMap, nil
+}
+
 // Sign the message with this node and aggregates with all child signatures (in structResponses)
 // Also aggregates the child bitmasks
-func generateSignature(ps pairing.Suite, t *onet.TreeNodeInstance, publics []kyber.Point, private kyber.Scalar,
-	responses map[network.ServerIdentityID]*Response, msg []byte, ok bool) (kyber.Point, *cosi.Mask, error) {
-
-	if t == nil {
-		return nil, nil, fmt.Errorf("TreeNodeInstance should not be nil, but is")
-	} else if responses == nil {
-		return nil, nil, fmt.Errorf("StructResponse should not be nil, but is")
-	} else if publics == nil {
-		return nil, nil, fmt.Errorf("publics should not be nil, but is")
-	} else if msg == nil {
-		return nil, nil, fmt.Errorf("msg should not be nil, but is")
-	}
+func (p *BlsFtCosi) generateSignature(responses ResponseMap, ok bool) (kyber.Point, *cosi.Mask, error) {
+	publics := p.Roster().Publics()
 
 	//generate personal mask
 	var personalMask *cosi.Mask
 	if ok {
-		personalMask, _ = cosi.NewMask(ps.(cosi.Suite), publics, t.Public())
+		personalMask, _ = cosi.NewMask(p.suite.(cosi.Suite), publics, p.Public())
 	} else {
-		personalMask, _ = cosi.NewMask(ps.(cosi.Suite), publics, nil)
+		personalMask, _ = cosi.NewMask(p.suite.(cosi.Suite), publics, nil)
 	}
 
 	// generate personal signature and append to other sigs
-	personalSig, err := bls.Sign(ps, private, msg)
+	personalSig, err := bls.Sign(p.suite, p.Private(), p.Msg)
 	if err != nil {
 		return nil, nil, err
 	}
 
 	// fill the map with the Root signature
-	responses[t.ServerIdentity().ID] = &Response{
+	responses[p.TreeNode().ID] = &Response{
 		Mask:      personalMask.Mask(),
 		Signature: personalSig,
 	}
 
 	// Aggregate all signatures
-	response, err := makeAggregateResponse(ps, publics, responses)
+	response, err := makeAggregateResponse(p.suite, publics, responses)
 	if err != nil {
-		log.Lvl3(t.ServerIdentity().Address, "failed to create aggregate signature")
+		log.Lvlf3("%v failed to create aggregate signature", p.ServerIdentity())
 		return nil, nil, err
 	}
 
 	//create final aggregated mask
-	finalMask, err := cosi.NewMask(ps.(cosi.Suite), publics, nil)
+	finalMask, err := cosi.NewMask(p.suite.(cosi.Suite), publics, nil)
 	if err != nil {
 		return nil, nil, err
 	}
@@ -325,11 +427,11 @@ func generateSignature(ps pairing.Suite, t *onet.TreeNodeInstance, publics []kyb
 		return nil, nil, err
 	}
 
-	finalSignature, err := response.Signature.Point(ps)
+	finalSignature, err := response.Signature.Point(p.suite)
 	if err != nil {
 		return nil, nil, err
 	}
-	log.Lvl3(t.ServerIdentity().Address, "is done aggregating signatures with total of", len(responses), "signatures")
+	log.Lvlf3("%v is done aggregating signatures with total of %d signatures", p.ServerIdentity(), len(responses))
 
 	return finalSignature, finalMask, err
 }
